driver-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-json-logger
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
deprecated